+++ src/main/java/com/omgdendi/blps/repository/UserRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface UserRepo extends CrudRepository<UserEntity, Long> {
+public interface UserRepo extends CrudRepository<UserEntity, Integer> {
     Optional<UserEntity> findByUsername(String username);
 }
